style(GameParticipantCards): Class 이름 및, method 이름 변경
- GameParticipant가 아닌 Cards를 책임지는 객체임이 드러나도록 이름 변경
- MaxScore 넘었는지 확인 메서드 isBust로 변경
- 해당하는 유저가 Bust인지 확인하는 test 작성
